# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for JWT authentication using Gin and Gorm.",
        "title": "JWT Auth API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
        "/auth/check-token": {
            "post": {
                "description": "检查当前 Token 的有效时间",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "检查 Token 是否快过期",
                "parameters": [
                    {
                        "description": "Token",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CheckTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token status",
                        "schema": {
                            "$ref": "#/definitions/controllers.ApiResult-float64"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/controllers.ApiResult-string"
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/controllers.ApiResult-string"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "用户通过用户名和密码进行登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.LoginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token",
                        "schema": {
                            "$ref": "#/definitions/controllers.ApiResult-string"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/controllers.ApiResult-string"
                        }
                    },
                    "401": {
                        "description": "Invalid password",
                        "schema": {
                            "$ref": "#/definitions/controllers.ApiResult-string"
                        }
                    }
                }
            }
        },
        "/auth/refresh-token": {
            "post": {
                "description": "使用刷新 Token 获取新的访问 Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "刷新 Token",
                "parameters": [
                    {
                        "description": "刷新信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.RefreshTokenInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New token",
                        "schema": {
                            "$ref": "#/definitions/controllers.ApiResult-string"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/controllers.ApiResult-string"
                        }
                    },
                    "401": {
                        "description": "Invalid refresh token",
                        "schema": {
                            "$ref": "#/definitions/controllers.ApiResult-string"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "用户通过用户名和密码进行注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.RegisterInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User registered successfully",
                        "schema": {
                            "$ref": "#/definitions/controllers.ApiResult-string"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/controllers.ApiResult-string"
                        }
                    },
                    "500": {
                        "description": "Failed to register user",
                        "schema": {
                            "$ref": "#/definitions/controllers.ApiResult-string"
                        }
                    }
                }
            }
        },
        "/user/profile": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取当前登录用户的信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取用户信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户 userID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户信息",
                        "schema": {
                            "$ref": "#/definitions/controllers.ApiResult-models_User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.ApiData-float64": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "改为 interface{} 类型"
                },
                "expires_in": {
                    "description": "以秒为单位的剩余时间",
                    "type": "integer"
                }
            }
        },
        "controllers.ApiData-models_User": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "改为 interface{} 类型"
                },
                "expires_in": {
                    "description": "以秒为单位的剩余时间",
                    "type": "integer"
                }
            }
        },
        "controllers.ApiData-string": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "改为 interface{} 类型"
                },
                "expires_in": {
                    "description": "以秒为单位的剩余时间",
                    "type": "integer"
                }
            }
        },
        "controllers.ApiResult-float64": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/controllers.ApiData-float64"
                },
                "error_details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "controllers.ApiResult-models_User": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/controllers.ApiData-models_User"
                },
                "error_details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "controllers.ApiResult-string": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/controllers.ApiData-string"
                },
                "error_details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "controllers.CheckTokenRequest": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "your_jwt_token_here"
                }
            }
        },
        "controllers.LoginInput": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "controllers.RefreshTokenInput": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "controllers.RegisterInput": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "JWT Token (format: Bearer \u003ctoken\u003e)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}